Used a more sophisticated deletion method
If, for example, you had cloned a git repository into the htdocs
directory, but you wanted to start over, the `rm -rf htdocs/*` will
not do the trick.  It would only delete files that are not hidden,
i.e. it would not delete the `.git` directory.  When you moved to the
`git clone` step, it would fail, saying there is already a repository
in the destination folder.