Dropped /all/ the tables before creating new ones
Primarily the purpose of this change is to drop all the tables before
trying to create new ones.  The reason is simple: You need to remove
the ones which have foreign key constraints first before you can
remove the ones which they reference.  If you do that as you create
them, then you end up with an illogical flow of creating linked tables
before you create the tables they link to.